+++ tracker/platforms/shared/main.c
@@ -9,10 +9,12 @@
 #include "corelib_mainloop.h"
 #include "app.h"
 #include "common.h"
+#include "keyboard_layout.h"
 
 int main(int argv, char** args) {
   // Load settings
   settingsLoad();
+  updateKeyboardLayout();
 
   if (gfxSetup(&appSettings.screenWidth, &appSettings.screenHeight) != 0) return 1;