+++ tracker/src/corelib/corelib_gfx.h
@@ -1,6 +1,8 @@
 #ifndef __CORELIB_GFX_H__
 #define __CORELIB_GFX_H__
 
+#include <stdint.h>
+
 /**
  * @brief Initialize graphics system. If screenWidth and screenHeight are not NULL or don't contain zeros,
  * function should read the screen resolution from these pointers. Otherwise, the function would detect screen resolution
@@ -56,5 +58,28 @@ void gfxCursor(int x, int y, int w);
 void gfxRect(int x, int y, int w, int h);
 void gfxPrint(int x, int y, const char* text);
 void gfxPrintf(int x, int y, const char* format, ...);
+
+/**
+ * @brief Draw 8-bit grayscale bitmap at character position
+ * 
+ * @param bitmap Grayscale bitmap data (row by row, 0=background, 255=foreground)
+ * @param col Column position in characters
+ * @param row Row position in characters
+ */
+void gfxDrawCharBitmap(uint8_t* bitmap, int col, int row);
+
+/**
+ * @brief Get character width in pixels
+ * 
+ * @return int Character width
+ */
+int gfxGetCharWidth(void);
+
+/**
+ * @brief Get character height in pixels
+ * 
+ * @return int Character height
+ */
+int gfxGetCharHeight(void);
 
 #endif